[topo] extract namespace to its own module
clarifies typed/untyped cache boundary better i think
